Refactor zeroization internals into separate modules
