+++ src/zeroizing_alloc.rs
@@ -104,8 +104,7 @@ where
         // and not yet deallocated SAFETY: `ptr` is at least `layout.align()`
         // byte aligned and this is a power of two
         unsafe {
-            self.zeroizer
-                .zeroize_mem(ptr, layout.size());
+            self.zeroizer.zeroize_mem(ptr, layout.size());
         }
         // SAFETY: caller must uphold the safety contract of `GlobalAlloc::dealloc`.
         unsafe { self.backend_alloc.dealloc(ptr, layout) }
@@ -164,8 +163,7 @@ where
         // SAFETY: `ptr` is at least `layout.align()` byte aligned and this is a power
         // of two
         unsafe {
-            self.zeroizer
-                .zeroize_mem(ptr.as_ptr(), layout.size());
+            self.zeroizer.zeroize_mem(ptr.as_ptr(), layout.size());
         }
         // SAFETY: caller must uphold the safety contract of `Allocator::deallocate`
         unsafe { self.backend_alloc.deallocate(ptr, layout) }
